# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 神农溪分布式多 Agent GIS API
  summary: Versioned workflow, resource, and internal Agent contracts
  description: The Master owns workflow routes; Publisher owns read-only tile/download routes. Internal
    routes are reachable only on the Compose network.
  version: 1.0.0
paths:
  /api/v1/tasks:
    post:
      tags:
      - Master public
      summary: Create Task
      description: Accept a Chinese ecological-monitoring request without blocking on analysis.
      operationId: createTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskAcceptedResponse'
        '422':
          description: Structured error without secret or provider payload fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Structured error without secret or provider payload fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/tasks/{task_id}:
    get:
      tags:
      - Master public
      summary: Get Task
      description: Reconstruct durable task, plan, step, error, and artifact state.
      operationId: getTask
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Structured error without secret or provider payload fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Structured error without secret or provider payload fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/tasks/{task_id}/events:
    get:
      tags:
      - Master public
      summary: Stream Task Events
      operationId: streamTaskEvents
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            text/event-stream:
              schema:
                type: string
        '404':
          description: Structured error without secret or provider payload fields.
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Structured error without secret or provider payload fields.
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      description: SSE stream. Each event id is the durable monotonic sequence and each data field is
        one TaskEvent JSON object. Last-Event-ID requests replay.
  /api/v1/tasks/{task_id}/retry:
    post:
      tags:
      - Master public
      summary: Retry Task
      description: Create an immutable new attempt from a validated safe checkpoint.
      operationId: retryTask
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetryAcceptedResponse'
        '404':
          description: Structured error without secret or provider payload fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Structured error without secret or provider payload fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Structured error without secret or provider payload fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/tiles/{task_id}/{artifact_type}/{z}/{x}/{y}.png:
    get:
      tags:
      - Publisher resources
      summary: Get Artifact Tile
      operationId: getArtifactTile
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      - name: artifact_type
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/TileArtifactType'
      - name: z
        in: path
        required: true
        schema:
          type: integer
          maximum: 24
          minimum: 0
          title: Z
      - name: x
        in: path
        required: true
        schema:
          type: integer
          minimum: 0
          title: X
      - name: y
        in: path
        required: true
        schema:
          type: integer
          minimum: 0
          title: Y
      responses:
        '200':
          description: Successful Response
          content:
            image/png:
              schema:
                type: string
        '404':
          description: Structured error without secret or provider payload fields.
          content:
            image/png:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Structured error without secret or provider payload fields.
          content:
            image/png:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Structured error without secret or provider payload fields.
          content:
            image/png:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/tasks/{task_id}/artifacts/{artifact_id}/download:
    get:
      tags:
      - Publisher resources
      summary: Download Artifact
      operationId: downloadArtifact
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      - name: artifact_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Artifact Id
      responses:
        '200':
          description: Successful Response
          content:
            application/octet-stream:
              schema:
                type: string
        '404':
          description: Structured error without secret or provider payload fields.
          content:
            application/octet-stream:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Structured error without secret or provider payload fields.
          content:
            application/octet-stream:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Structured error without secret or provider payload fields.
          content:
            application/octet-stream:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /internal/v1/data/prepare:
    post:
      tags:
      - Internal Agent
      summary: Prepare Data
      operationId: prepareData
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataPrepareCommand'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataPrepareResult'
        '409':
          description: Structured error without secret or provider payload fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Structured error without secret or provider payload fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Structured error without secret or provider payload fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /internal/v1/analysis/run:
    post:
      tags:
      - Internal Agent
      summary: Run Analysis
      operationId: runAnalysis
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalysisRunCommand'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisRunResult'
        '409':
          description: Structured error without secret or provider payload fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Structured error without secret or provider payload fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Structured error without secret or provider payload fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /internal/v1/quality/evaluate:
    post:
      tags:
      - Internal Agent
      summary: Evaluate Quality
      operationId: evaluateQuality
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QualityEvaluateCommand'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityEvaluateResult'
        '409':
          description: Structured error without secret or provider payload fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Structured error without secret or provider payload fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Structured error without secret or provider payload fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /internal/v1/publisher/publish:
    post:
      tags:
      - Internal Agent
      summary: Publish Results
      operationId: publishResults
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublisherPublishCommand'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublisherPublishResult'
        '409':
          description: Structured error without secret or provider payload fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Structured error without secret or provider payload fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Structured error without secret or provider payload fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    AnalysisRunCommand:
      properties:
        schema_version:
          type: string
          const: '1.0'
          title: Schema Version
          default: '1.0'
        task_id:
          type: string
          format: uuid
          title: Task Id
        step_id:
          type: string
          pattern: ^[a-z][a-z0-9_]{0,63}$
          title: Step Id
        attempt:
          type: integer
          minimum: 1.0
          title: Attempt
        correlation_id:
          type: string
          format: uuid
          title: Correlation Id
        inputs:
          items:
            $ref: '#/components/schemas/DataAssetRef'
          type: array
          title: Inputs
      additionalProperties: false
      type: object
      required:
      - task_id
      - step_id
      - attempt
      - correlation_id
      - inputs
      title: AnalysisRunCommand
    AnalysisRunResult:
      properties:
        schema_version:
          type: string
          const: '1.0'
          title: Schema Version
          default: '1.0'
        task_id:
          type: string
          format: uuid
          title: Task Id
        step_id:
          type: string
          pattern: ^[a-z][a-z0-9_]{0,63}$
          title: Step Id
        attempt:
          type: integer
          minimum: 1.0
          title: Attempt
        correlation_id:
          type: string
          format: uuid
          title: Correlation Id
        artifacts:
          items:
            $ref: '#/components/schemas/ArtifactRef'
          type: array
          title: Artifacts
        statistics:
          $ref: '#/components/schemas/AreaStatistics'
      additionalProperties: false
      type: object
      required:
      - task_id
      - step_id
      - attempt
      - correlation_id
      - artifacts
      - statistics
      title: AnalysisRunResult
...
    CreateTaskRequest:
      properties:
        schema_version:
          type: string
          const: '1.0'
          title: Schema Version
          default: '1.0'
        query:
          type: string
          maxLength: 2000
          minLength: 1
          title: Query
      additionalProperties: false
      type: object
      required:
      - query
      title: CreateTaskRequest
...
    DataPrepareCommand:
      properties:
        schema_version:
          type: string
          const: '1.0'
          title: Schema Version
          default: '1.0'
        task_id:
          type: string
          format: uuid
          title: Task Id
        step_id:
          type: string
          pattern: ^[a-z][a-z0-9_]{0,63}$
          title: Step Id
        attempt:
          type: integer
          minimum: 1.0
          title: Attempt
        correlation_id:
          type: string
          format: uuid
          title: Correlation Id
        dataset_ids:
          items:
            $ref: '#/components/schemas/LogicalDatasetId'
          type: array
          title: Dataset Ids
      additionalProperties: false
      type: object
      required:
      - task_id
      - step_id
      - attempt
      - correlation_id
      - dataset_ids
      title: DataPrepareCommand
    DataPrepareResult:
      properties:
        schema_version:
          type: string
          const: '1.0'
          title: Schema Version
          default: '1.0'
        task_id:
          type: string
          format: uuid
          title: Task Id
        step_id:
          type: string
          pattern: ^[a-z][a-z0-9_]{0,63}$
          title: Step Id
        attempt:
          type: integer
          minimum: 1.0
          title: Attempt
        correlation_id:
          type: string
          format: uuid
          title: Correlation Id
        assets:
          items:
            $ref: '#/components/schemas/DataAssetRef'
          type: array
          title: Assets
      additionalProperties: false
      type: object
      required:
      - task_id
      - step_id
      - attempt
      - correlation_id
      - assets
      title: DataPrepareResult
...
    ErrorResponse:
      properties:
        schema_version:
          type: string
          const: '1.0'
          title: Schema Version
          default: '1.0'
        error:
          $ref: '#/components/schemas/StructuredError'
      additionalProperties: false
      type: object
      required:
      - error
      title: ErrorResponse
    ExecutionPlan:
      properties:
        schema_version:
          type: string
          const: '1.0'
          title: Schema Version
          default: '1.0'
        plan_id:
          type: string
          format: uuid
          title: Plan Id
        task_id:
          type: string
          format: uuid
          title: Task Id
        source:
          $ref: '#/components/schemas/PlanSource'
        created_at:
          type: string
          format: date-time
          title: Created At
        model_call:
          anyOf:
          - $ref: '#/components/schemas/ModelCallRecord'
          - type: 'null'
        steps:
          items:
            $ref: '#/components/schemas/PlanStep'
          type: array
          title: Steps
      additionalProperties: false
      type: object
      required:
      - plan_id
      - task_id
      - source
      - created_at
      - steps
      title: ExecutionPlan
...
    LogicalDatasetId:
      type: string
      enum:
      - watershed
      - before_red
      - before_nir
      - after_red
      - after_nir
      title: LogicalDatasetId
...
    PublishedResource:
      properties:
        schema_version:
          type: string
          const: '1.0'
          title: Schema Version
          default: '1.0'
        artifact_id:
          type: string
          format: uuid
          title: Artifact Id
        tile_template:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Tile Template
        download_path:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Download Path
      additionalProperties: false
      type: object
      required:
      - artifact_id
      title: PublishedResource
    PublisherPublishCommand:
      properties:
        schema_version:
          type: string
          const: '1.0'
          title: Schema Version
          default: '1.0'
        task_id:
          type: string
          format: uuid
          title: Task Id
        step_id:
          type: string
          pattern: ^[a-z][a-z0-9_]{0,63}$
          title: Step Id
        attempt:
          type: integer
          minimum: 1.0
          title: Attempt
        correlation_id:
          type: string
          format: uuid
          title: Correlation Id
        artifacts:
          items:
            $ref: '#/components/schemas/ArtifactRef'
          type: array
          title: Artifacts
        quality:
          $ref: '#/components/schemas/QualityMetrics'
      additionalProperties: false
      type: object
      required:
      - task_id
      - step_id
      - attempt
      - correlation_id
      - artifacts
      - quality
      title: PublisherPublishCommand
    PublisherPublishResult:
      properties:
        schema_version:
          type: string
          const: '1.0'
          title: Schema Version
          default: '1.0'
        task_id:
          type: string
          format: uuid
          title: Task Id
        step_id:
          type: string
          pattern: ^[a-z][a-z0-9_]{0,63}$
          title: Step Id
        attempt:
          type: integer
          minimum: 1.0
          title: Attempt
        correlation_id:
          type: string
          format: uuid
          title: Correlation Id
        resources:
          items:
            $ref: '#/components/schemas/PublishedResource'
          type: array
          title: Resources
        report:
          $ref: '#/components/schemas/ArtifactRef'
      additionalProperties: false
      type: object
      required:
      - task_id
      - step_id
      - attempt
      - correlation_id
      - resources
      - report
      title: PublisherPublishResult
    QualityEvaluateCommand:
      properties:
        schema_version:
          type: string
          const: '1.0'
          title: Schema Version
          default: '1.0'
        task_id:
          type: string
          format: uuid
          title: Task Id
        step_id:
          type: string
          pattern: ^[a-z][a-z0-9_]{0,63}$
          title: Step Id
        attempt:
          type: integer
          minimum: 1.0
          title: Attempt
        correlation_id:
          type: string
          format: uuid
          title: Correlation Id
        artifacts:
          items:
            $ref: '#/components/schemas/ArtifactRef'
          type: array
          title: Artifacts
        analysis_elapsed_ms:
          type: integer
          minimum: 0.0
          title: Analysis Elapsed Ms
      additionalProperties: false
      type: object
      required:
      - task_id
      - step_id
      - attempt
      - correlation_id
      - artifacts
      - analysis_elapsed_ms
      title: QualityEvaluateCommand
    QualityEvaluateResult:
      properties:
        schema_version:
          type: string
          const: '1.0'
          title: Schema Version
          default: '1.0'
        task_id:
          type: string
          format: uuid
          title: Task Id
        step_id:
          type: string
          pattern: ^[a-z][a-z0-9_]{0,63}$
          title: Step Id
        attempt:
          type: integer
          minimum: 1.0
          title: Attempt
        correlation_id:
          type: string
          format: uuid
          title: Correlation Id
        metrics:
          $ref: '#/components/schemas/QualityMetrics'
        artifact:
          $ref: '#/components/schemas/ArtifactRef'
      additionalProperties: false
      type: object
      required:
      - task_id
      - step_id
      - attempt
      - correlation_id
      - metrics
      - artifact
      title: QualityEvaluateResult
...
    RetryAcceptedResponse:
      properties:
        schema_version:
          type: string
          const: '1.0'
          title: Schema Version
          default: '1.0'
        task_id:
          type: string
          format: uuid
          title: Task Id
        attempt:
          type: integer
          minimum: 2.0
          title: Attempt
        status:
          $ref: '#/components/schemas/TaskStatus'
        accepted_at:
          type: string
          format: date-time
          title: Accepted At
      additionalProperties: false
      type: object
      required:
      - task_id
      - attempt
      - status
      - accepted_at
      title: RetryAcceptedResponse
...
    StructuredError:
      properties:
        schema_version:
          type: string
          const: '1.0'
          title: Schema Version
          default: '1.0'
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          maxLength: 2000
          minLength: 1
          title: Message
        retryable:
          type: boolean
          title: Retryable
        details:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          type: array
          title: Details
          default: []
      additionalProperties: false
      type: object
      required:
      - code
      - message
      - retryable
      title: StructuredError
    TaskAcceptedResponse:
      properties:
        schema_version:
          type: string
          const: '1.0'
          title: Schema Version
          default: '1.0'
        task_id:
          type: string
          format: uuid
          title: Task Id
        status:
          $ref: '#/components/schemas/TaskStatus'
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - task_id
      - status
      - created_at
      title: TaskAcceptedResponse
...
    TaskResponse:
      properties:
        schema_version:
          type: string
          const: '1.0'
          title: Schema Version
          default: '1.0'
        task_id:
          type: string
          format: uuid
          title: Task Id
        query:
          type: string
          maxLength: 2000
          minLength: 1
          title: Query
        status:
          $ref: '#/components/schemas/TaskStatus'
        progress:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: Progress
        current_attempt:
          type: integer
          minimum: 1.0
          title: Current Attempt
        correlation_id:
          type: string
          format: uuid
          title: Correlation Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        plan:
          anyOf:
          - $ref: '#/components/schemas/ExecutionPlan'
          - type: 'null'
        steps:
          items:
            $ref: '#/components/schemas/TaskStep'
          type: array
          title: Steps
          default: []
        artifacts:
          items:
            $ref: '#/components/schemas/ArtifactRef'
          type: array
          title: Artifacts
          default: []
        last_error:
          anyOf:
          - $ref: '#/components/schemas/StructuredError'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - task_id
      - query
      - status
      - progress
      - current_attempt
      - correlation_id
      - created_at
      - updated_at
      title: TaskResponse
    TaskStatus:
      type: string
      enum:
      - PENDING
      - PLANNING
      - DATA_PREPARING
      - ANALYZING
      - QUALITY_CHECKING
      - PUBLISHING
      - COMPLETED
      - FAILED
      title: TaskStatus
      description: Durable workflow states exposed by the public API.
    TaskStep:
      properties:
        schema_version:
          type: string
          const: '1.0'
          title: Schema Version
          default: '1.0'
        step_id:
          type: string
          pattern: ^[a-z][a-z0-9_]{0,63}$
          title: Step Id
        kind:
          $ref: '#/components/schemas/PlanStepKind'
        agent:
          $ref: '#/components/schemas/AgentName'
        attempt:
          type: integer
          minimum: 1.0
          title: Attempt
        status:
          $ref: '#/components/schemas/StepStatus'
        progress:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: Progress
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
        elapsed_ms:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Elapsed Ms
        error:
          anyOf:
          - $ref: '#/components/schemas/StructuredError'
          - type: 'null'
        artifacts:
          items:
            $ref: '#/components/schemas/ArtifactRef'
          type: array
          title: Artifacts
          default: []
      additionalProperties: false
      type: object
      required:
      - step_id
      - kind
      - agent
      - attempt
      - status
      - progress
      title: TaskStep
    TileArtifactType:
      type: string
      enum:
      - NDVI_BEFORE
      - NDVI_AFTER
      - NDVI_DIFFERENCE
      - CHANGE_CLASSIFICATION
      title: TileArtifactType